Minor refactoring on finance support.

+++ Cod.Platform.Finance/AccountBalance.cs
@@ -2,10 +2,10 @@
 {
     public class AccountBalance
     {
-        public double Total { get; set; }
+        public long Total { get; set; }
 
-        public double Frozen { get; set; }
+        public long Frozen { get; set; }
 
-        public double Available { get; set; }
+        public long Available { get; set; }
     }
 }
